Exception Reorganization & System Views Unavailable
Added a new system unavailable exception handling & a html page for it.
