+++ cassanova/api/bootstrap.py
@@ -3,9 +3,11 @@ from starlette.staticfiles import StaticFiles
 
 from cassanova.config.app_config import APPConfig
 from cassanova.consts.app_routers import APPConsts
-from cassanova.exceptions.cluster_not_available import add_cluster_unavailable_exceptions
-from cassanova.exceptions.default_handler import add_default_exceptions
-from cassanova.exceptions.not_found_handler import add_notfound_exceptions
+from cassanova.exceptions.route_handlers.cluster_unavailable_handler import add_cluster_unavailable_exceptions
+from cassanova.exceptions.route_handlers.default_handler import add_default_exceptions
+from cassanova.exceptions.route_handlers.not_found_handler import add_notfound_exceptions
+from cassanova.exceptions.route_handlers.system_views_unavailable_handler import \
+    add_system_views_unavailable_exception_handler
 
 
 def bootstrap_app(app: FastAPI, app_config: APPConfig):
@@ -28,3 +30,4 @@ def __add_exception_handlers(app: FastAPI):
     add_default_exceptions(app)
     add_notfound_exceptions(app)
     add_cluster_unavailable_exceptions(app)
+    add_system_views_unavailable_exception_handler(app)
