+++ cassanova/exceptions/route_handlers/default_handler.py
