+++ cassanova/exceptions/system_views_unavailable.py
@@ -1,0 +1,3 @@
+class SystemViewsUnavailableException(Exception):
+    def __init__(self, message: str):
+        self.message = message
