+++ cassanova/web/templates/exceptions/system-views-unavailable.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8"/>
+    <title>System Views Not Available</title>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap" rel="stylesheet"/>
+    <link href="/static/styles/exceptions.css" rel="stylesheet"/>
+    <link href="/static/styles/main.css" rel="stylesheet"/>
+    <script src="/static/scripts/sidebar.js" defer></script>
+</head>
+<body>
+<!-- Sidebar -->
+<aside class="sidebar collapsed">
+    <div class="sidebar-header">
+        <a href="/" class="sidebar-brand">
+            <img src="/static/icons/logo/cassanova-logo.png" alt="Cassanova Logo" class="sidebar-logo"/>
+            <span class="sidebar-title">Cassanova</span>
+        </a>
+        <button class="sidebar-toggle" aria-label="Toggle Menu" aria-expanded="false">
+            <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#e0e0e0" stroke-width="2"
+                 stroke-linecap="round" stroke-linejoin="round">
+                <line x1="3" y1="6" x2="21" y2="6"/>
+                <line x1="3" y1="12" x2="21" y2="12"/>
+                <line x1="3" y1="18" x2="21" y2="18"/>
+            </svg>
+        </button>
+    </div>
+    <nav class="sidebar-nav">
+        <a href="/" class="sidebar-link">
+            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
+                 stroke="currentColor" class="icon">
+                <path stroke-linecap="round" stroke-linejoin="round"
+                      d="M3 12l2-2m0 0l7-7 7 7M5 10v10a1 1 0 001 1h3m10-11l2 2m-2-2v10a1 1 0 01-1 1h-3"/>
+            </svg>
+            <span class="label">Home</span>
+        </a>
+        <a href="{{ monitoring_url }}" class="sidebar-link">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" stroke-width="1.5" fill="none"
+                 stroke="currentColor" class="icon">
+                <path stroke-linecap="round" stroke-linejoin="round"
+                      d="M2.036 12.322a1.012 1.012 0 0 1 0-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178Z"/>
+                <circle cx="12" cy="12" r="3"/>
+            </svg>
+            <span class="label">Monitoring</span>
+        </a>
+        <a href="/tools" class="sidebar-link">
+            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
+                 stroke="currentColor" class="icon">
+                <path stroke-linecap="round" stroke-linejoin="round"
+                      d="M11.42 15.17 17.25 21A2.652 2.652 0 0 0 21 17.25l-5.877-5.877M11.42 15.17l2.496-3.03c.317-.384.74-.626 1.208-.766M11.42 15.17l-4.655 5.653a2.548 2.548 0 1 1-3.586-3.586l6.837-5.63m5.108-.233c.55-.164 1.163-.188 1.743-.14a4.5 4.5 0 0 0 4.486-6.336l-3.276 3.277a3.004 3.004 0 0 1-2.25-2.25l3.276-3.276a4.5 4.5 0 0 0-6.336 4.486c.091 1.076-.071 2.264-.904 2.95l-.102.085m-1.745 1.437L5.909 7.5H4.5L2.25 3.75l1.5-1.5L7.5 4.5v1.409l4.26 4.26"/>
+            </svg>
+            <span class="label">Tools</span>
+        </a>
+    </nav>
+</aside>
+
+<div class="starfield"></div>
+
+<div class="cluster-down-container" role="alert" aria-live="assertive" aria-atomic="true">
+    <h1 style="color: #4fc3f7; font-weight: 700; font-size: 2.5rem; margin-bottom: 1rem;">System Views Not Available</h1>
+    <p style="color: #e0e0e0; font-size: 1.2rem; max-width: 600px; margin: 0 auto 1.5rem;">
+        The <code>system_views.settings</code> table is not accessible or this technology does not support it.
+    </p>
+    <p style="color: #888; font-style: italic;">
+        Please check your cluster or use other means to view configuration.
+    </p>
+    <img class="logo" src="/static/icons/logo/cassanova-logo.png" alt="Cassanova Icon" style="margin-top: 2rem; width: 80px; opacity: 0.7;"/>
+</div>
+</body>
+</html>
